requirements.txt
# exact dependency versions used in development
oset==0.1.3
PyYAML==3.11
